feat: support inline answer collection definition inside activity templates
